### 2018-01-19 #540 @whedon
[PRE REVIEW]: py-opc: operate the Alphasense OPC-N2 from a raspberry pi or other popular microcontrollers/microcomputers
**Submitting author:** @dhhagan (<a href="http://orcid.org/0000-0001-5111-4671">David Henry Hagan</a>)
**Repository:** <a href="https://github.com/dhhagan/py-opc" target ="_blank">https://github.com/dhhagan/py-opc</a>
**Version:** 1.4.1
