### 2018-01-19 #540 @dhhagan
Hey @arfon , thanks for the PR (now merged). The research purpose is that it allows researchers to operate the particle counter from any microcontroller, rather than a PC, and is used by many research organizations (EPA, UNEP, MIT, Leeds, etc) to do aerosol research. Maybe a bit more information is given in a pre-submission inquiry I posted last month (https://github.com/openjournals/joss/issues/362).

Let me know if this doesn't fit and I can withdraw.
