### 2018-01-20 #540 @arfon
> Hey @arfon , thanks for the PR (now merged). The research purpose is that it allows researchers to operate the particle counter from any microcontroller, rather than a PC, and is used by many research organizations (EPA, UNEP, MIT, Leeds, etc) to do aerosol research. Maybe a bit more information is given in a pre-submission inquiry I posted last month (openjournals/joss#362).

Thanks @dhhagan - I'd forgotten that you'd already enquired. We should be able to proceed here.

While we wait for an editor to volunteer (and then find a reviewer) would you mind fleshing out your paper with a little more detail? Our goal is to have short papers that are understandable by a non-specialist audience. Right now I don't think this paper is quite there.
