### 2018-01-24 2084b56
rewrite of paper.md
