### 2018-01-24 #540 @dhhagan
Okay. I gave a rewrite a stab. I read a bunch of the accepted papers and tried to model it based on those, so hopefully it is okay now.
