### 2018-01-26 #540 @Kevin-Mattheus-Moerman
👋 @rajeshxsankaran @jared711 @skgrange would you be interested in reviewing [this project](https://github.com/dhhagan/py-opc) for [the Journal of Open Source Software](http://joss.theoj.org/), the [review process](http://joss.theoj.org/about#reviewer_guidelines) focuses largely on the software and a short the [paper](https://github.com/openjournals/joss-papers/blob/joss.00540/joss.00540/10.21105.joss.00540.pdf). Let me know if you are available to help so I can open a review issue for this submission.
