### 2018-01-26 #540 @Kevin-Mattheus-Moerman
@dhhagan thanks for submitting this paper to JOSS. I've started looking for reviewers. If you could suggest suitable reviewers that would be very helpful. Also your work appears to rely on hardware. Will reviewers require this hardware for review? Can some aspects of the code be reviewed without the hardware?
