### 2018-01-26 #540 @dhhagan
Hey @Kevin-Mattheus-Moerman: it does rely on hardware - I guess it depends on exactly what the review entails, but it would probably work better with the hardware. I suppose, depending on where the reviewers are located, I could lend a unit for review? What has been done in the past with code that relies on hardware?
