### 2018-03-01 #540 @dhhagan
Yup. Absolutely. Do you need a Pi as well?
On Thu, Mar 1, 2018 at 4:56 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @dhhagan <https://github.com/dhhagan> could you drop off the device at
> the MIT Media Lab ​E14-274G (my office)? I'll then pass it on to
> @amjaeger17 <https://github.com/amjaeger17> who can help with review.
>
