+++ README.md
@@ -10,21 +10,30 @@ Python library for operating the Alphasense OPC-N2 Optical Particle Counter usin
 
 ## Dependencies
 
-One of:
+One of the following, depending on whether you use GPIO pins or a SPI-USB adapter:
 
-  1. [`py-spidev`](https://github.com/doceme/py-spidev)
-  1. [`py-usbiss`](https://github.com/dancingquanta/py-usbiss)
+  1. [`py-spidev`](https://github.com/doceme/py-spidev) - for those using GPIO pins
+  1. [`pyusbiss`](https://github.com/dancingquanta/pyusbiss) - for those using a SPI-USB adapter (python3+ only)
 
 
 ## Installation
 
 For use on the Raspberry Pi (or any other linux device?), install via pip:
 
-    >>> pip install py-opc [--upgrade]
+    $ pip install py-opc [--upgrade]
+
+As `pyusbiss` is not yet available through a package manager, you must download from source (if you are using the SPI-USB adapter only). This can be done as follows:
+
+    $ pip install git+https://github.com/DancingQuanta/pyusbiss.git
+
+If you are using the GPIO pins to communicate with the OPC-N2, you must download the requirement `py-spidev` as follows:
+
+    $ pip install git+https://github.com/doceme/py-spidev.git
+
 
 ## License
 
-  This library is licensed under the MIT license.
+  This library is licensed under the MIT license. The full text of the license can be found in this repository at LICENSE.txt.
 
 ## Documentation
 
@@ -32,10 +41,16 @@ For use on the Raspberry Pi (or any other linux device?), install via pip:
 
   You can also build the documentation by navigating to the `docs` directory and issuing the command:
 
-    >>> make html
+    $ make html
 
 
 ## Sample Script / Getting Started
+
+To quickly get up and running, follow one of the two examples:
+
+### GPIO-connected OPC-N2
+
+Use if you are using the GPIO pins in conjunction with `py-spidev`
 
     import spidev
     import opc
@@ -59,3 +74,26 @@ For use on the Raspberry Pi (or any other linux device?), install via pip:
 
     # Turn the opc OFF
     alphasense.off()
+
+### SPI-USB Adapter with OPC-N2
+
+Use this approach if you have connected your RPi to the OPC-N2 via a SPI-USB adapter.
+
+**NOTE**: Currently, this method is only supported on python3+ due to limitations in the `pyusbiss` library.
+
+    import usbiss
+    import opc
+
+    # Build the connector
+    spi = usbiss.USBISS("/dev/ttyACM0", "spi", spi_mode=2, freq=500000)
+
+    alpha = opc.OPCN2(spi)
+
+    # Turn on the device
+    alpha.on()
+
+    # Read the histogram
+    alpha.histogram()
+
+    # Turn the device off
+    alpha.off()
