### 2018-03-08 0dc481c
Updated instructions for pyusbiss 0.2.0
