### 2018-03-08 cd2b2b3
Renamed usbiss.USBISS to usbiss.spi.SPI
