### 2018-03-08 86e97a4
Updated README.md for pyusbiss 0.2.0
