### 2018-04-18 e696b73
added support for retrying a serial connection and got rid of support for firmware versions <= 7
