### 2018-04-18 9c4dc5d
added support for logging through the python logging module
