### 2018-04-18 22de358
edited manual tests
