### 2018-05-07 #540 @dhhagan
Hey @Kevin-Mattheus-Moerman I dropped off the device with you, right? Could I come and grab it sometime this week - it needs to go out into the field soon. Thanks!
