### 2018-05-07 #540 @Kevin-Mattheus-Moerman
Yes you can. Does Friday work?
