### 2018-05-11 4c7790e
Fixed a bug when setting laser power
