### 2018-06-15 #540 @Kevin-Mattheus-Moerman
The device is still here at the MIT media lab. My colleague @amjaeger17 can unfortunately no longer do this review. I will ask another colleague to take a look but if he cannot then perhaps we are at a dead end for now. Apologies for the delay this has causes with this submission.
