### 2018-06-19 #782 @Kevin-Mattheus-Moerman
Hi @dhhagan your references field in the paper appears empty. Can you check this out? If you make changes to the paper and want to re-generate the pdf here you can also use:
`@whedon generate pdf`
Let me know if you have questions.
