### 2018-06-20 #782 @dhhagan
@Kevin-Mattheus-Moerman Interesting - I thought they were supposed to be somehow pulled from the bib file? I tried copying over from some accepted repositories; is there an easy way to pull them in this way?
