### 2018-06-20 #782 @arfon
@dhhagan - you need to cite them directly for them to be included in the compilation. (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
