### 2018-06-21 920f7f0
updated bib file for joss submission
