### 2018-06-21 #782 @dhhagan
@arfon Is there a way to locally generate the pdf to validate the references show up? It seems like there must be an error with the BibTex syntax, but I can't seem to figure out what's going on. Is there any output from the `@whedon generate pdf` command?
