### 2018-06-21 #782 @arfon
@dhhagan does this help? https://github.com/greatfireball/docker_whedon_translate4joss

[This is the command](https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L109-L132) that Whedon executes to compile the PDF.
