### 2018-06-25 357a306
added sleep-wait after several commands
