### 2018-06-25 #782 @Kevin-Mattheus-Moerman
@dhhagan in relation to the release/version number. This is no problem. We can update the version number at a later stage.
