### 2018-06-26 76bfd70
Added await function to alpha
