### 2018-06-27 00d7f08
updated readme to reflect sleeping habits
