+++ setup.py
@@ -5,7 +5,7 @@
     Written originally by David H Hagan
 '''
 
-VERSION = '1.6.1'
+VERSION = '1.6.0'
 
 try:
     from setuptools import setup
