### 2018-06-27 #782 @tshu
@dhhagan Everything looks good. Once I can check the pdf for fixed references, my review will be complete.
