### 2018-06-28 #782 @dhhagan
Thanks @tshu I'll try and figure out what's wrong with the `papers.bib` file a bit later today...not sure what's going on there.
