### 2018-06-28 #782 @Kevin-Mattheus-Moerman
@dhagan did you try building the paper locally to trouble shoot this? perhaps remove the references and add a sigle bib file entry at a time to try to figure out what is going on.
