### 2018-06-28 #782 @arfon
The issue here is that you're not actually citing your references in the `paper.md` file which means they aren't included by Pandoc. Can you cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
