### 2018-06-28 #782 @danielskatz
\hijack-thread-on

@arfon - can we provide some guidance on this in our docs?  (which maybe just means pointing to the link you provided as part of a "how to write a JOSS paper and include references" section)

\hijack-thread-off
