### 2018-06-28 #782 @Kevin-Mattheus-Moerman
So @dhhagan it looks like you need to cite them inline in the text e.g. like:
```latex
Bla bla @py_usbiss and bla bla @py_spidev, bladibla @alphasense_opcn2.
```
Please also specifically cite the Crilley et al. article that is in review, this way it also appears in the reference list.
