### 2018-06-28 #782 @arfon
> @arfon - can we provide some guidance on this in our docs? (which maybe just means pointing to the link you provided as part of a "how to write a JOSS paper and include references" section)

@danielskatz - we already having this in the example `paper.md` here: http://joss.theoj.org/about#author_guidelines but perhaps we could make this more obvious?

```Citations to entries in paper.bib should be in
[rMarkdown](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
format.```
