+++ paper/paper.bib
@@ -21,3 +21,10 @@ year = 2016,
 url = {https://github.com/DancingQuanta/pyusbiss},
 urldate = {2018-01-19}
 }
+
+@online{Crilley:2018,
+author = {Crilley, L. R., Shaw, M., Pound, R., Kramer, L. J., Price, R., Young, S., Lewis, A. C., and Pope, F. D},
+title = {Evaluation of a low-cost optical particle counter (Alphasense OPC-N2) for ambient air monitoring},
+year = 2018,
+doi = {https://doi.org/10.5194/amt-11-709-2018},
+journal = {Atmospheric Measurement Techniques},}
