### 2018-06-28 c901971
cleaned up references
