### 2018-06-28 #782 @dhhagan
@arfon @Kevin-Mattheus-Moerman Thanks for the help - I didn't notice the individual citations throughout the abstract the first time around - I assumed it was just including the line `bibliography: paper.bib`. Regardless, it should be good now. Thanks!
