### 2018-06-28 #782 @Kevin-Mattheus-Moerman
@dhhagan great it looks like we are nearly there. @dhhagan
1) About the citation for `doceme 2015`, the author name field is a github name. Perhaps this is okay but it would be nice if we could replace this by a human name. I cannot see the name of this author either so perhaps we are forced to do it that way. @doceme How can we best cite you here?

@arfon
1) Do you know if the citations `(doceme 2015, Tolmie (2016))` render okay? Seems inconsistent to have `name date` and `name (date)` like that, it seems to mix two formats? Or is that normal for our reference format, like one is rendered as a url accessed on a particular data and the other is an article citation?
2) For some reason all references in this paper have the date in them twice. Is this normal? Other papers do not seem to have it like that. I did notice that there is a `year` field as well as a `urldate` field in 3 of the 4 references but the 4th reference does not have that...
