+++ paper/paper.bib
@@ -7,7 +7,7 @@ urldate = {2018-01-19}
 }
 
 @online{py_spidev,
-author = {doceme},
+author = {Stephen Caudle},
 title = {py-spidev},
 year = 2015,
 url = {https://github.com/doceme/py-spidev},
