### 2018-06-28 #782 @arfon
> Do you know if the citations (doceme 2015, Tolmie (2016)) render okay? Seems inconsistent to have name date and name (date) like that, it seems to mix two formats? Or is that normal for our reference format, like one is rendered as a url accessed on a particular data and the other is an article citation?

This can be fixed by citing the papers like this `[@py_spidev; @py_usbiss]`

> For some reason all references in this paper have the date in them twice. Is this normal? Other papers do not seem to have it like that. I did notice that there is a year field as well as a urldate field in 3 of the 4 references but the 4th reference does not have that...

I think this is because these are type `online`. The `Crilley:2018` reference should be `article` type which will remove the extra date.
