+++ paper/paper.bib
@@ -22,7 +22,7 @@ url = {https://github.com/DancingQuanta/pyusbiss},
 urldate = {2018-01-19}
 }
 
-@online{Crilley:2018,
+@article{Crilley:2018,
 author = {Crilley, L. R., Shaw, M., Pound, R., Kramer, L. J., Price, R., Young, S., Lewis, A. C., and Pope, F. D},
 title = {Evaluation of a low-cost optical particle counter (Alphasense OPC-N2) for ambient air monitoring},
 year = 2018,
