### 2018-06-28 #782 @tshu
@dhhagan The references have been fixed, and I have checked off all review points. The device itself has been left with @Kevin-Mattheus-Moerman for pickup.

As an additional note, the readme mentions that only python3 is supported for USB-SPI, though installation instructions through pip are for python2. It's a minor thing, but perhaps a one liner about how the installation can be run with pip3 would be good. It definitely caught me (hence the opened issue.)
