### 2018-06-28 #782 @Kevin-Mattheus-Moerman
@dhhagan thanks for amending the paper. Two tiny things and we are done here.

1) Consider implementing the suggestion by @tshu in relation to the readme.
2) @doceme created a DOI for citing his work, see here: https://zenodo.org/record/1299863#.WzUvXhwh1G8
So please could you replace the citation with:
```latex
@misc{stephen_caudle_2018_1299863,
  author       = {Stephen Caudle},
  title        = {doceme/py-spidev: Version 3.3},
  month        = jun,
  year         = 2018,
  doi          = {10.5281/zenodo.1299863},
  url          = {https://doi.org/10.5281/zenodo.1299863}
}
```
Once you've done these steps, please regenerate the paper here and archive a version of the final reviewed software in a service like [Zenodo](https://zenodo.org/), and share the DOI of this archive here. @arfon can then proceed to formally accept this submission (see also: https://guides.github.com/activities/citable-code/). N.B. please inform @arfon if the release/version number has changed since review (or as part of the DOI minting) so he can update it for you.

Thanks @tshu for the review work here! :tada:
