### 2018-06-28 cd50f84
finalized joss release
