+++ paper/paper.bib
@@ -14,6 +14,15 @@ url = {https://github.com/doceme/py-spidev},
 urldate = {2018-01-19}
 }
 
+@misc{stephen_caudle_2018_1299863,
+author = {Stephen Caudle},
+title = {doceme/py-spidev: Version 3.3},
+month = jun,
+year = 2018,
+doi = {10.5281/zenodo.1299863},
+url = {https://doi.org/10.5281/zenodo.1299863}
+}
+
 @online{py_usbiss,
 author = {Andrew Tolmie},
 title = {py-usbiss},
