### 2018-06-28 #782 @dhhagan
Should be good to go:

Zenodo DOI: DOI: 10.5281/zenodo.1299205
Final reviewed version: v1.6.0
